+++ widget/android/AndroidAlerts.cpp
@@ -6,6 +6,7 @@
 #include "AndroidAlerts.h"
 #include "mozilla/java/GeckoRuntimeWrappers.h"
 #include "mozilla/java/WebNotificationWrappers.h"
+#include "mozilla/java/WebNotificationActionWrappers.h"
 #include "nsIPrincipal.h"
 #include "nsIURI.h"
 
@@ -102,6 +103,24 @@ AndroidAlerts::ShowAlert(nsIAlertNotification* aAlert,
   rv = aAlert->GetVibrate(vibrate);
   NS_ENSURE_SUCCESS(rv, NS_OK);
 
+  nsTArray<RefPtr<nsIAlertAction>> nsActions;
+  MOZ_TRY(aAlert->GetActions(nsActions));
+  jni::ObjectArray::LocalRef actions =
+      jni::ObjectArray::New(nsActions.Length());
+  size_t index = 0;
+  for (auto& nsAction : nsActions) {
+    nsAutoString name;
+    MOZ_TRY(nsAction->GetAction(name));
+
+    nsAutoString title;
+    MOZ_TRY(nsAction->GetTitle(title));
+
+    java::WebNotificationAction::LocalRef action =
+        java::WebNotificationAction::New(name, title);
+    actions->SetElement(index, action);
+    ++index;
+  }
+
   if (!sNotificationMap) {
     sNotificationMap = new NotificationMap();
   } else if (Maybe<AndroidNotificationTuple> tuple =
@@ -113,7 +132,7 @@ AndroidAlerts::ShowAlert(nsIAlertNotification* aAlert,
 
   java::WebNotification::LocalRef notification = notification->New(
       title, name, cookie, text, imageUrl, dir, lang, requireInteraction, spec,
-      silent, privateBrowsing, jni::IntArray::From(vibrate));
+      silent, privateBrowsing, jni::IntArray::From(vibrate), actions);
   AndroidNotificationTuple tuple{
       .mObserver = aAlertListener,
       .mAlert = aAlert,
@@ -164,7 +183,7 @@ NS_IMETHODIMP AndroidAlerts::Teardown() {
 NS_IMETHODIMP AndroidAlerts::PbmTeardown() { return NS_ERROR_NOT_IMPLEMENTED; }
 
 void AndroidAlerts::NotifyListener(const nsAString& aName, const char* aTopic,
-                                   const char16_t* aCookie) {
+                                   Maybe<nsString> aAction) {
   if (!sNotificationMap) {
     return;
   }
@@ -175,7 +194,11 @@ void AndroidAlerts::NotifyListener(const nsAString& aName, const char* aTopic,
   }
 
   if (tuple->mObserver) {
-    tuple->mObserver->Observe(nullptr, aTopic, nullptr);
+    nsCOMPtr<nsIAlertAction> action;
+    if (aAction) {
+      tuple->mAlert->GetAction(*aAction, getter_AddRefs(action));
+    }
+    tuple->mObserver->Observe(action, aTopic, nullptr);
   }
 
   if ("alertfinished"_ns.Equals(aTopic)) {